Add 3D plotting capabilities to `averages.py`
Calling `averages.py` with `--3d` will generate 3D plots instead of 2D
ones. Specifically, the plots will use the same colouring, but the 3D
version also provides height context for the magnitude of the
coefficients.

This is especially useful for comparing the order of magnitudes of
differences between single coefficients.